Refactor: 직원 상세정보 페이지 컴포넌트 구조 변경
