+++ src/components/Employee/AddMemberModal.tsx
@@ -1,9 +1,10 @@
-import React from "react";
+import React, { useState } from "react";
 import { Button } from "antd";
 import { UserAddOutlined } from "@ant-design/icons";
 import CustomForm from "../common/CustomForm";
-import AddMemberForm from "./AddMemberForm";
 import styled from "styled-components";
+import MemberForm from "./MemberForm";
+import MemberProfile from "./MemberProfile";
 
 const SumbitBtn = styled.div`
   display: flex;
@@ -13,6 +14,7 @@ const SumbitBtn = styled.div`
 export default function AddMemberModal({ onCancel }: { onCancel: () => void }) {
   const Form = CustomForm.Form;
   const [form] = Form.useForm();
+  const [isEditMode, setIsEditMode] = useState(true);
 
   const handleSubmit = () => {
     onCancel();
@@ -27,7 +29,8 @@ export default function AddMemberModal({ onCancel }: { onCancel: () => void }) {
       }}
       form={form}
     >
-      <AddMemberForm form={form} />
+      <MemberProfile isEditMode={isEditMode} />
+      <MemberForm isEditMode={isEditMode} />
       <SumbitBtn>
         <Button icon={<UserAddOutlined />} htmlType="submit" type="primary">
           Add
